UPDATE: Use u32 for sheet indices in function signatures (#23)
That removes problem with casting i32 into unsigned number. At the
moment we assume that u32 to usize cast is safe for sheet index.

+++ packages/calc/equalto_calc/src/functions/lookup_and_reference.rs
@@ -698,7 +698,7 @@ impl Model {
                         let sheet_index;
                         if let Some(name) = sheet_name {
                             match self.get_sheet_index_by_name(&name) {
-                                Some(i) => sheet_index = i as i32,
+                                Some(i) => sheet_index = i,
                                 None => {
                                     return CalcResult::Error {
                                         error: Error::REF,
@@ -728,7 +728,7 @@ impl Model {
                         let sheet_index;
                         if let Some(name) = sheet_name {
                             match self.get_sheet_index_by_name(&name) {
-                                Some(i) => sheet_index = i as i32,
+                                Some(i) => sheet_index = i,
                                 None => {
                                     return CalcResult::Error {
                                         error: Error::REF,
